initial commit
Initial commit.

This is basically the philtroyGodot3 branch. functionalitywise, but with a bunch of refactoring and new graphics.

Ideally, this shouldn´t be released to public yet, but I wanted to create the repository so I can start using GitHub´s project management tools to create bug reports and tasks.

Once all graphics are updated, and editor is able to view and edit simple dialogue, I will release as alpha 1 (should be done by end of this week)