Editor WIP
Editor WIP.

If you want an idea on how the GUI works:

§ to start editor
down arrow to close editor

choose dialogue file from the list on the far left, navigate back and forth the dialogue tree by clicking on the reply nodes.

ctrl+LMB on replies to edit
ctrl+enter to save edit

buttons and checkboxes do nothing right now, just placeholders to get an idea of the final editor.

disclaimer: this does not actually save dialogue edits to file yet, though it updates texts in the editor nodes. There are also UI bugs I need to address, and short keys are not final.

But I´m very happy with the general workflow, so time to wrap things up before the end of the week :)

A last note, the first version of the editor will only be for plain dialogue text edit. Events and triggers will follow later. Working in small steps ;P