Saving a node in the editor now saves to the editor cache
still have a weird bug when going back and forth in the dialogue tree, which is proving annying as f-ck (I´ll fix it eventually:) Meanwhile, I setup the editor to actually be able to save edited nodes to the editor cache. From their it´s easy peasy to save back to file (old file should be backed up)

Baby steps! :)