pick up objects
Created global func remove_from_scene()
Now, if you leftclick an object it gets removed from the scene, and if you use another object on it, it will give you a remark.

This is obviously very barebones.  Notification should be shown when picking up an object, object should be put in inventory, maybe an animated effect when picking it up. Low priority right now.,