WIP Editor
Editor now checks for dialogue variables and flags, activates checkboxes for those when editing replies.

Also, fixed a bug where reply couldn´t be edited if it had exit true.

The box that pops up to the right of the boxes is just for debugging.

These are small steps toward making the editor truly usable for game building. For that, we need to be able to edit those variables and flags in the editor. Plan is to use the debug window as the foundation for the UI we´ll need for that.

Little by little :)