speech balloon refactor and more
Speech balloon system was refactored to be more general purpose and dynamic. In the process added create_timer and wait_and_execute functions to global.gd.

A lot of general refactoring of code to make more easily readable.  A bunch of deprecated code and nodes were removed, especially from game.gd, which is much leaner as a result.

ApeVERSE.exe updated to Godot 3.4.4
Readme.md updated tp note that game needs to be loaded once in Godot first to generate necessary import file